src/stt_service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==1.10.13
aiohttp==3.9.1
redis==5.0.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
httpx==0.27.2  # Pin to 0.27.2 to avoid 'proxies' argument error
livekit==1.0.0
pyjwt==2.8.0
webrtcvad-wheels==2.0.14  # For streaming VAD, replaces silero-vad 
deepgram-sdk==3.2.0
websockets==15.0.1